requirements.txt
asgiref==3.5.0
Django==3.2.12
django-ajax-datatable==4.4.2
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.1.0
